tech(api): move some session-management security pre-handlers in certification scope
