+++ api/src/certification/session-management/application/usecases/checkUserIsMemberOfCertificationCenterSession.js
@@ -1,5 +1,5 @@
-import * as sessionRepository from '../../../certification/session-management/infrastructure/repositories/session-management-repository.js';
-import * as certificationCourseRepository from '../../../certification/shared/infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
+import * as sessionRepository from '../../infrastructure/repositories/session-management-repository.js';
 
 const execute = async function ({
   userId,
