+++ api/tests/shared/unit/application/security-pre-handlers_test.js
@@ -1290,247 +1290,6 @@ describe('Shared | Unit | Application | SecurityPreHandlers', function () {
     });
   });
 
-  describe('#checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId', function () {
-    context('Successful case', function () {
-      it('should authorize access to resource when the user is a member of the organization center', async function () {
-        // given
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().resolves(true),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId(
-            {
-              auth: { credentials: { accessToken: 'valid.access.token', userId: 123 } },
-              params: { certificationCourseId: 7 },
-            },
-            hFake,
-            {
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.source).to.be.true;
-      });
-    });
-
-    context('Error cases', function () {
-      it('should forbid resource access when user was not previously authenticated', async function () {
-        // given
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub(),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId(
-            { auth: { credentials: {} }, params: { certificationCourseId: 5678 } },
-            hFake,
-            {
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when user is not a member of the organization center', async function () {
-        // given
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().resolves(false),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId(
-            {
-              auth: { credentials: { accessToken: 'valid.access.token', userId: 1 } },
-              params: { certificationCourseId: 5678 },
-            },
-            hFake,
-            {
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when an error is thrown by use case', async function () {
-        // given
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().rejects(new Error('Some error')),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId(
-            {
-              auth: { credentials: { accessToken: 'valid.access.token', userId: 1 } },
-              params: { certificationCourseId: 5678 },
-            },
-            hFake,
-            {
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-    });
-  });
-
-  describe('#checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId', function () {
-    context('Successful case', function () {
-      it('should authorize access to resource when the user is a member of the organization center', async function () {
-        // given
-        const certificationCourseId = 7;
-
-        const certificationIssueReportRepositoryStub = {
-          get: sinon.stub().withArgs(certificationCourseId).resolves({ certificationCourseId }),
-        };
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().resolves(true),
-        };
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId(
-            {
-              auth: { credentials: { accessToken: 'valid.access.token', userId: 123 } },
-              params: { id: 666 },
-            },
-            hFake,
-            {
-              certificationIssueReportRepository: certificationIssueReportRepositoryStub,
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.source).to.be.true;
-      });
-    });
-
-    context('Error cases', function () {
-      it('should forbid resource access when user was not previously authenticated', async function () {
-        // given
-        const certificationIssueReportRepositoryStub = {
-          get: sinon.stub(),
-        };
-
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub(),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId(
-            { auth: { credentials: {} }, params: { id: 5678 } },
-            hFake,
-            {
-              certificationIssueReportRepository: certificationIssueReportRepositoryStub,
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when user is not a member of the organization center', async function () {
-        // given
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().resolves(false),
-        };
-        const certificationIssueReportRepositoryStub = {
-          get: sinon.stub().resolves({ certificationCourseId: 7 }),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId(
-            { auth: { credentials: { accessToken: 'valid.access.token', userId: 1 } }, params: { id: 666 } },
-            hFake,
-            {
-              certificationIssueReportRepository: certificationIssueReportRepositoryStub,
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when an error is thrown by use case', async function () {
-        // given
-        const certificationIssueReportRepositoryStub = {
-          get: sinon.stub().resolves({ certificationCourseId: 7 }),
-        };
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub().rejects(new Error('Some error')),
-        };
-
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId(
-            { auth: { credentials: { accessToken: 'valid.access.token', userId: 1 } }, params: { id: 666 } },
-            hFake,
-            {
-              certificationIssueReportRepository: certificationIssueReportRepositoryStub,
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-
-      it('should forbid resource access when an error is thrown by repo', async function () {
-        // given
-        const certificationIssueReportRepositoryStub = {
-          get: sinon.stub().rejects(new Error('Some error')),
-        };
-        const checkUserIsMemberOfCertificationCenterSessionUsecaseStub = {
-          execute: sinon.stub(),
-        };
-        // when
-        const response =
-          await securityPreHandlers.checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId(
-            { auth: { credentials: { accessToken: 'valid.access.token', userId: 1 } }, params: { id: 666 } },
-            hFake,
-            {
-              certificationIssueReportRepository: certificationIssueReportRepositoryStub,
-              checkUserIsMemberOfCertificationCenterSessionUsecase:
-                checkUserIsMemberOfCertificationCenterSessionUsecaseStub,
-            },
-          );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-    });
-  });
-
   describe('#checkUserOwnsCertificationCourse', function () {
     context('Successful case', function () {
       it('should authorize access to resource when the user owns the certification course', async function () {
